+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 Release 2.0.2:
  - `vclib-openid`: Add response modes for query and fragment, i.e. Wallet may return the authentication response in query params or as fragment params on a SIOPv2 call
+ - `vclib-openid`: Create fresh challenges for every SIOPv2 request
 
 Release 2.0.1:
  - `vclib-openid`: Remove `OidcSiopProtocol`, replace with `OidcSiopVerifier` and `OidcSiopWallet`, remove `AuthenticationResponse` and `AuthenticationRequest` holder classes